--- add-on/src/lib/helpers.js
@@ -1,1 +1,0 @@
-export const dropSlash = url => url.replace(/\/$/, '')
